Market Basket Analysis when items are in a single column
So far, the tutorials that I have seen online use a data frame that already has each  row representing a transaction having the respective items. However, my data frame has all the items in a single column called 'model_id' meaning each row was added when a purchase was made. Also, the apriori's rules in Python are not easy to view. Hence, my code has the initial and last few steps to make it easier to use apriori in Python.